use mask operator to avoid branch

+++ source/Lib/TLibCommon/TComRom.h
@@ -128,7 +128,7 @@ extern const int16_t g_chromaFilter[8][NTAPS_CHROMA]; ///< Chroma filter taps
 // Scanning order & context mapping table
 // ====================================================================================================================
 
-extern const uint32_t g_groupIdx[32];
+extern const uint8_t g_groupIdx[32];
 extern const uint32_t g_minInGroup[10];
 
 extern const uint8_t g_goRiceRange[5];      //!< maximum value coded with Rice codes
